+++ scripts/workflows/provision-darwin.sh
@@ -28,8 +28,8 @@ if [ "$E2E_TEST" = "tests-dfx/certificate.bash" ]; then
      brew install mitmproxy
 fi
 if [ "$E2E_TEST" = "tests-dfx/deps.bash" ]; then
-     cargo install cargo-binstall@1.6.9
-     cargo binstall -y ic-wasm
+     cargo install cargo-binstall@1.6.9 --locked
+     cargo binstall -y ic-wasm --locked
 fi
 
 if [ "$E2E_TEST" = "tests-icx-asset/icx-asset.bash" ]; then