+++ e2e/tests-dfx/build_rust.bash
@@ -27,7 +27,7 @@ teardown() {
 @test "build with vulnerabilities in rust dependencies emits a warning" {
   assert_command rustup default stable
   assert_command rustup target add wasm32-unknown-unknown
-  assert_command cargo install cargo-audit
+  assert_command cargo install cargo-audit --locked
   assert_command cargo audit --version
   install_asset vulnerable_rust_deps
   dfx_start