+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 Report upgrade compatibility warnings for Motoko, such as deleted stable variables, in addition to compatibility errors.
 
+### feat: PocketIC state
+
+`dfx start --pocketic` no longer requires `--clean`, and can persist replica state between runs.
+
+# 0.23.0
+
 ### feat: Add canister snapshots
 
 The new `dfx canister snapshot` command can be used to create, apply, and delete snapshots of stopped canisters.
@@ -36,7 +42,6 @@ This incorporates the following executed proposals:
 
 - [131473](https://dashboard.internetcomputer.org/proposal/131473)
 
-# 0.22.1
 
 ## Dependencies